Add files via upload
This is a python program to generate a random password for your accounts so that all your accounts are safe. You can enter different parameter to customize your password. The level of this project is level 1 as a beginner with just basic of knowledge of python can also make this project. So what are you waiting for customize this project with your own ideas make your personal password generator.